Move to getColumns for better flexibility
