+++ src/platform/plugins/shared/data/public/search/session/sessions_mgmt/components/table/actions/popover_actions.tsx
@@ -31,6 +31,7 @@ interface PopoverActionItemsProps {
   api: SearchSessionsMgmtAPI;
   onActionComplete: OnActionComplete;
   core: CoreStart;
+  allowedActions?: UISession['actions'];
 }
 
 export const PopoverActionsMenu = ({
@@ -38,6 +39,7 @@ export const PopoverActionsMenu = ({
   onActionComplete,
   session,
   core,
+  allowedActions,
 }: PopoverActionItemsProps) => {
   const [isPopoverOpen, setPopover] = useState(false);
 
@@ -66,7 +68,11 @@ export const PopoverActionsMenu = ({
     </EuiToolTip>
   );
 
-  const actions = session.actions || [];
+  const actions =
+    session.actions?.filter((action) => {
+      if (!allowedActions) return true;
+      return allowedActions.includes(action);
+    }) || [];
   // Generic set of actions - up to the API to return what is available
   const items = actions.reduce((itemSet, actionType) => {
     const actionDef = getAction(api, actionType, session, core);
