+++ src/platform/plugins/shared/data/public/search/session/sessions_mgmt/components/table/columns/index.ts
@@ -7,5 +7,7 @@
  * License v3.0 only", or the "Server Side Public License, v 1".
  */
 
-export * as columns from './columns';
-export { SearchSessionsMgmtTable } from './table';
+export * from './app_id';
+export * from './name';
+export * from './actions';
+export * from './status';
