+++ src/platform/plugins/shared/data/public/search/session/sessions_mgmt/components/table/columns/status.tsx
@@ -1,0 +1,25 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the "Elastic License
+ * 2.0", the "GNU Affero General Public License v3.0 only", and the "Server Side
+ * Public License v 1"; you may not use this file except in compliance with, at
+ * your election, the "Elastic License 2.0", the "GNU Affero General Public
+ * License v3.0 only", or the "Server Side Public License, v 1".
+ */
+
+import { i18n } from '@kbn/i18n';
+import React from 'react';
+import { EuiBasicTableColumn } from '@elastic/eui';
+import { UISession } from '../../../types';
+import { StatusIndicator } from '../../status';
+
+export const statusColumn = (timezone: string): EuiBasicTableColumn<UISession> => ({
+  field: 'status',
+  name: i18n.translate('data.mgmt.searchSessions.table.headerStatus', {
+    defaultMessage: 'Status',
+  }),
+  sortable: true,
+  render: (statusType: UISession['status'], session) => (
+    <StatusIndicator session={session} timezone={timezone} />
+  ),
+});
